pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "seeding"
version = "0.1.0"
description = "Image editor with YOLOv8 detection"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "PyQt5>=5.15",
    "qt-material>=2.14",
    "numpy>=1.23",
    "opencv-python>=4.7.0",
    "Pillow>=9.4.0",
    "PyMuPDF>=1.22.0",
    "ultralytics>=8.0.0",
    "reportlab>=3.6.0",
]

[project.scripts]
seeding = "seeding:main"
